+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Command `grunt build:admin` has been replaced by `npm run build:back-office-client`
 - Command `grunt unit-tests` has been replaced by `npm run test:client`
 - Command `grunt test-e2e` has been replaced by `npm run e2e`
+- Command `grunt compile-js` has been removed
+- Command `grunt dist` has been replaced by `npm run build`
 - Command `grunt` has been replaced by `npm run watch -- --with-source-maps`
 
 # 7.0.0 / 2020-11-19